migrated profile methods to user
